Generate document history annex: https://github.com/metanorma/metanorma-ogc/issues/152
