+++ Gemfile.devel
@@ -1,0 +1,1 @@
+gem "metanorma-standoc", git: "https://github.com/metanorma/metanorma-standoc", branch: "features/pub-history"
