+++ lib/isodoc/ogc/presentation_xml_convert.rb
@@ -1,6 +1,6 @@
 require_relative "init"
-require "uuidtools"
 require_relative "../../relaton/render/general"
+require_relative "presentation_sections"
 
 module IsoDoc
   module Ogc
@@ -9,8 +9,6 @@ module IsoDoc
         @libdir = File.dirname(__FILE__)
         super
       end
-
-      def middle_title(docxml); end
 
       def convert1(docxml, filename, dir)
         info docxml, nil
@@ -21,115 +19,10 @@ module IsoDoc
         File.read(File.join(@libdir, "html", "rouge.css"))
       end
 
-      def preface_rearrange(doc)
-        super
-        insert_preface_sections(doc)
-      end
-
-      def insert_preface_sections(doc)
-        preface_insert(doc.at(ns("//preface/clause" \
-                                 "[@type = 'executivesummary']")),
-                       doc.at(ns("//preface/abstract")), doc)
-        preface_insert(doc.at(ns("//preface//submitters")),
-                       submit_orgs_append_pt(doc), doc)
-        insert_submitting_orgs(doc)
-        preface_insert(doc.at(ns("//preface/clause[@type = 'security']")),
-                       submit_orgs_append_pt(doc), doc)
-        insert_keywords(doc)
-      end
-
-      def preface_init_insert_pt(docxml)
-        docxml.at(ns("//preface")) ||
-          docxml.at(ns("//sections"))
-            .add_previous_sibling("<preface> </preface>").first
-      end
-
-      def preface_insert(clause, after, docxml)
-        clause or return
-        clause.remove
-        if after then after.next = clause
-        else
-          preface_init_insert_pt(docxml)&.children&.first
-            &.add_previous_sibling(clause)
-        end
-      end
-
-      def submit_orgs_append_pt(docxml)
-        docxml.at(ns("//foreword")) ||
-          docxml.at(ns("//preface/clause[@type = 'keywords']")) ||
-          docxml.at(ns("//preface/clause[@type = 'executivesummary']")) ||
-          docxml.at(ns("//preface/abstract"))
-      end
-
-      def insert_submitting_orgs(docxml)
-        orgs = docxml.xpath(ns(submittingorgs_path))
-          .each_with_object([]) { |org, m| m << org.text }
-        orgs.empty? and return
-        if a = submit_orgs_append_pt(docxml)
-          a.next = submitting_orgs_clause(orgs)
-        else
-          preface_init_insert_pt(docxml)&.children&.first
-            &.add_previous_sibling(submitting_orgs_clause(orgs))
-        end
-      end
-
-      def submitting_orgs_clause(orgs)
-        <<~SUBMITTING
-          <clause id="_#{UUIDTools::UUID.random_create}" type="submitting_orgs">
-          <title>Submitting Organizations</title>
-          <p>The following organizations submitted this Document to the Open Geospatial Consortium (OGC):</p>
-           <ul>#{orgs.map { |m| "<li>#{m}</li>" }.join("\n")}</ul>
-           </clause>
-        SUBMITTING
-      end
-
-      def keyword_clause(kwords)
-        <<~KEYWORDS
-          <clause id="_#{UUIDTools::UUID.random_create}" type="keywords">
-          <title>Keywords</title>
-          <p>The following are keywords to be used by search engines and document catalogues.</p>
-          <p>#{kwords.join(', ')}</p></clause>
-        KEYWORDS
-      end
-
-      def insert_keywords(docxml)
-        kw = @meta.get[:keywords]
-        kw.empty? and return
-        if abstract =
-             docxml.at(ns("//preface/clause[@type = 'executivesummary']")) ||
-             docxml.at(ns("//preface/abstract"))
-          abstract.next = keyword_clause(kw)
-        else
-          preface_init_insert_pt(docxml)&.children&.first
-            &.add_previous_sibling(keyword_clause(kw))
-        end
-      end
-
       def example1(elem)
         lbl = @xrefs.anchor(elem["id"], :label, false) or return
         prefix_name(elem, block_delim, l10n("#{@i18n.example} #{lbl}"),
                     "name")
-      end
-
-      def annex1(elem)
-        lbl = @xrefs.anchor(elem["id"], :label)
-        t = elem.at(ns("./title")) and
-          t.children = "<strong>#{to_xml(t.children)}</strong>"
-        prefix_name(elem, "<br/>", lbl, "title")
-      end
-
-      def clause(docxml)
-        super
-        docxml.xpath(ns("//foreword | //preface/abstract | " \
-                        "//submitters | //introduction | //acknowledgements"))
-          .each { |f| clause1(f) }
-      end
-
-      def clause1(elem)
-        elem.name == "terms" && elem.parent.name == "annex" &&
-          elem.parent.xpath(ns("./terms")).size == 1 and return
-        elem.name == "clause" && elem["type"] == "toc" and return
-        super
       end
 
       def bibdata(docxml)
@@ -138,6 +31,88 @@ module IsoDoc
             a["type"] = "contributor"
         end
         super
+        dochistory_insert(docxml)
+      end
+
+      UPDATE_RELATIONS = <<~XPATH.freeze
+        //bibdata/relation[@type = 'updatedBy' or @type = 'merges' or @type = 'splits' or @type = 'hasDraft']/bibitem
+      XPATH
+
+      def dochistory_insert(docxml)
+        updates = docxml.xpath(ns(UPDATE_RELATIONS))
+        updates.empty? and return
+        fwd = annex_insert_point(docxml)
+        generate_dochistory(updates, fwd)
+      end
+
+      def annex_insert_point(docxml)
+        docxml.at(ns("//annex[last()]")) || docxml.at(ns("//sections"))
+      end
+
+      def generate_dochistory(updates, pref)
+        ret = updates.map { |u| generate_dochistory_row(u) }.flatten.join("\n")
+        pref.next = <<~XML
+          <annex id='_#{UUIDTools::UUID.random_create}' obligation='informative'>
+          <title>#{@i18n.dochistory}</title>
+          <table><thead>
+          <tr><th>Date</th><th>Release</th><th>Author</th><th>Paragraph Modified</th><th>Description</th></tr>
+          </thead><tbody>
+          #{ret}
+          </tbody></table></annex>
+        XML
+      end
+
+      def generate_dochistory_row(item)
+        e = item.at(ns("./edition")) || item.at(ns("./version/draft"))
+        date = dochistory_date(item)
+        c = dochistory_contributors(item)
+        l = dochistory_location(item)
+        desc = dochistory_description(item)
+        "<tr><td>#{date}</td><td>#{e&.text}</td><td>#{c}</td>" \
+          "<td>#{l}</td><td>#{desc}</td></tr>"
+      end
+
+      def dochistory_date(item)
+        d = item.at(ns("./date[@type = 'updated']")) ||
+          item.at(ns("./date[@type = 'published']")) ||
+          item.at(ns("./date[@type = 'issued']")) or return ""
+        d.text.strip
+      end
+
+      def dochistory_contributors(item)
+        item.xpath(ns("./contributor")).map do |c|
+          dochistory_contributor(c)
+        end.join(", ")
+      end
+
+      def dochistory_contributor(contrib)
+        ret = contrib.at("./organization/subdivision") ||
+          contrib.at("./organization/name") ||
+          contrib.at("./person/name/abbreviation") ||
+          contrib.at("./person/name/completename")
+        ret and return ret.text
+        format_personalname(contrib)
+      end
+
+      def format_personalname(contrib)
+        Relaton::Render::Ogc::General
+          .new(template: { book: "{{ creatornames }}" })
+          .render("<bibitem type='book'>#{contrib.to_xml}</bibitem>",
+                  embedded: true)
+      end
+
+      def dochistory_description(item)
+        d = item.at(ns("./amend/description")) or return ""
+        d.children.to_xml
+      end
+
+      def dochistory_location(item)
+        t = item.at(ns("./amend/location")) or return "All"
+        xpath = "./amend/location/locality | ./amend/location/localityStack"
+        r = eref_localities(item.xpath(ns(xpath)), nil, t)
+        r.sub!(/^, /, "")
+        r == @i18n.wholeoftext and r = "All"
+        r
       end
 
       def bibdata_i18n(bib)
@@ -203,7 +178,7 @@ module IsoDoc
       # else, use both ordinal, as prefix, and ids
       def biblio_ref_entry_code(ordinal, ids, _id, standard, datefn, _bib)
         standard and return "[#{ordinal}]<tab/>"
-        ret = (ids[:ordinal] || ids[:metanorma] || "[#{ordinal}]")
+        ret = ids[:ordinal] || ids[:metanorma] || "[#{ordinal}]"
         prefix_bracketed_ref("#{ret}#{datefn}")
       end
