+++ spec/isodoc/html_convert_spec.rb
@@ -1,388 +1,6 @@
 require "spec_helper"
 
-logoloc = File.expand_path(
-  File.join(File.dirname(__FILE__), "..", "..",
-            "lib", "isodoc", "ogc", "html"),
-)
-
 RSpec.describe IsoDoc::Ogc do
-  it "processes default metadata" do
-    csdc = IsoDoc::Ogc::HtmlConvert.new({})
-    input = <<~INPUT
-      <ogc-standard xmlns="https://standards.opengeospatial.org/document">
-      <bibdata type="standard">
-        <title language="en" format="text/plain">Main Title</title>
-        <uri>http://www.example.com</uri>
-        <uri type="html">http://www.example.com/html</uri>
-        <uri type="xml">http://www.example.com/xml</uri>
-        <uri type="pdf">http://www.example.com/pdf</uri>
-        <uri type="doc">http://www.example.com/doc</uri>
-        <docidentifier type="ogc-external">http://www.example2.com</docidentifier>
-        <docidentifier type="ogc-internal">1000</docidentifier>
-        <docnumber>1000</docnumber>
-        <abstract><p>This is a description.</p>
-        <quote>This is a <em>blockquote</em> within a description.</quote>
-        </abstract>
-        <date type="published">
-          <on>2002-01-01</on>
-        </date>
-        <date type="created">
-          <on>1999-01-01</on>
-        </date>
-        <date type="issued">
-          <on>2001-01-01</on>
-        </date>
-        <contributor>
-          <role type="author"/>
-          <organization>
-            <name>OGC</name>
-          </organization>
-        </contributor>
-        <contributor>
-          <role type="editor"/>
-          <person>
-            <name>
-              <completename>Fred Flintstone</completename>
-            </name>
-          </person>
-        </contributor>
-        <contributor>
-          <role type="author"/>
-          <person>
-            <name>
-              <forename>Barney</forename>
-              <surname>Rubble</surname>
-            </name>
-          </person>
-        </contributor>
-        <contributor>
-          <role type="contributor"/>
-          <person>
-            <name>
-              <forename>Pearl</forename>
-              <surname>Slaghoople</surname>
-            </name>
-          </person>
-        </contributor>
-        <contributor>
-          <role type="publisher"/>
-          <organization>
-            <name>OGC</name>
-          </organization>
-        </contributor>
-        <edition>2.0</edition>
-        <version>
-        <revision-date>2000-01-01</revision-date>
-        <draft>3.4</draft>
-      </version>
-        <language>en</language>
-        <script>Latn</script>
-        <status><stage>SWG Work</stage></status>
-        <copyright>
-          <from>2001</from>
-          <owner>
-            <organization>
-              <name>OGC</name>
-            </organization>
-          </owner>
-        </copyright>
-        <copyright>
-          <from>2001</from>
-          <owner>
-            <organization>
-              <name>ISO</name>
-            </organization>
-          </owner>
-        </copyright>
-        <copyright>
-          <from>2001</from>
-          <owner>
-            <organization>
-              <name>IEC</name>
-            </organization>
-          </owner>
-        </copyright>
-        <keyword>A</keyword>
-        <keyword>B</keyword>
-        <ext>
-        <doctype>standard</doctype>
-        <subdoctype>conceptual-model-and-encoding</subdoctype>
-        <editorialgroup>
-          <committee type="A">TC</committee>
-          <committee type="B">TC1</committee>
-          <subcommittee type="C" number="1">SC1</committee>
-          <workgroup type="D" number="2">WG1</committee>
-        </editorialgroup>
-        </ext>
-      </bibdata>
-      <sections/>
-      </ogc-standard>
-    INPUT
-
-    output = <<~"OUTPUT"
-      {:abstract=>"This is a description. This is a blockquote within a description.",
-      :accesseddate=>"XXX",
-      :adapteddate=>"XXX",
-      :agency=>"OGC",
-      :announceddate=>"XXX",
-      :authors=>["Barney Rubble"],
-      :circulateddate=>"XXX",
-      :confirmeddate=>"XXX",
-      :contributors=>["Pearl Slaghoople"],
-      :copieddate=>"XXX",
-      :copyright_holder=>"OGC, ISO, and IEC",
-      :correcteddate=>"XXX",
-      :createddate=>"1999-01-01",
-      :doc=>"http://www.example.com/doc",
-      :doclanguage=>"English",
-      :docnumber=>"1000",
-      :docnumeric=>"1000",
-      :docsubtype=>"Conceptual Model And Encoding",
-      :docsubtype_abbr=>"CME",
-      :doctitle=>"Main Title",
-      :doctype=>"Standard",
-      :doctype_abbr=>"IS",
-      :docyear=>"2001",
-      :draft=>"3.4",
-      :draftinfo=>" (draft 3.4, 2000-01-01)",
-      :edition=>"2.0",
-      :editors=>["Fred Flintstone"],
-      :externalid=>"http://www.example2.com",
-      :html=>"http://www.example.com/html",
-      :implementeddate=>"XXX",
-      :issueddate=>"2001-01-01",
-      :keywords=>["A", "B"],
-      :lang=>"en",
-      :logo_new=>"#{File.join(logoloc, 'logo.2021.svg')}",
-      :logo_old=>"#{File.join(logoloc, 'logo.png')}",
-      :logo_word=>"#{File.join(logoloc, 'logo.png')}",
-      :obsoleteddate=>"XXX",
-      :pdf=>"http://www.example.com/pdf",
-      :publisheddate=>"2002-01-01",
-      :publisher=>"OGC",
-      :receiveddate=>"XXX",
-      :revdate=>"2000-01-01",
-      :revdate_monthyear=>"January 2000",
-      :script=>"Latn",
-      :stable_untildate=>"XXX",
-      :stage=>"SWG Work",
-      :stage_display=>"SWG Work",
-      :stageabbr=>"SW",
-      :tc=>"TC",
-      :transmitteddate=>"XXX",
-      :unchangeddate=>"XXX",
-      :unpublished=>true,
-      :updateddate=>"XXX",
-      :url=>"http://www.example.com",
-      :vote_endeddate=>"XXX",
-      :vote_starteddate=>"XXX",
-      :xml=>"http://www.example.com/xml"}
-    OUTPUT
-    docxml, = csdc.convert_init(input, "test", true)
-    expect(htmlencode(metadata(csdc.info(docxml, nil)).to_s)
-      .gsub(", :", ",\n:")).to be_equivalent_to output
-  end
-
-  it "processes metadata with new logo" do
-    csdc = IsoDoc::Ogc::HtmlConvert.new({})
-    input = <<~INPUT
-      <ogc-standard xmlns="https://standards.opengeospatial.org/document">
-      <bibdata type="standard">
-        <title language="en" format="text/plain">Main Title</title>
-        <docidentifier type="ogc-external">http://www.example2.com</docidentifier>
-        <docidentifier type="ogc-internal">1000</docidentifier>
-        <docnumber>1000</docnumber>
-        <date type="published">
-          <on>2030-01-01</on>
-        </date>
-      </bibdata>
-      </ogc-standard>
-    INPUT
-    output = <<~OUTPUT
-      {:accesseddate=>"XXX",
-      :adapteddate=>"XXX",
-      :announceddate=>"XXX",
-      :circulateddate=>"XXX",
-      :confirmeddate=>"XXX",
-      :copieddate=>"XXX",
-      :copyright_holder=>"Open Geospatial Consortium",
-      :correcteddate=>"XXX",
-      :createddate=>"XXX",
-      :doclanguage=>"English",
-      :docnumber=>"1000",
-      :docnumeric=>"1000",
-      :doctitle=>"Main Title",
-      :externalid=>"http://www.example2.com",
-      :implementeddate=>"XXX",
-      :issueddate=>"XXX",
-      :lang=>"en",
-      :logo_new=>"#{File.join(logoloc, 'logo.2021.svg')}",
-      :logo_old=>"#{File.join(logoloc, 'logo.png')}",
-      :logo_word=>"#{File.join(logoloc, 'logo.2021.svg')}",
-      :obsoleteddate=>"XXX",
-      :publisheddate=>"2030-01-01",
-      :receiveddate=>"XXX",
-      :script=>"Latn",
-      :stable_untildate=>"XXX",
-      :transmitteddate=>"XXX",
-      :unchangeddate=>"XXX",
-      :unpublished=>true,
-      :updateddate=>"XXX",
-      :vote_endeddate=>"XXX",
-      :vote_starteddate=>"XXX"}
-    OUTPUT
-    docxml, = csdc.convert_init(input, "test", true)
-    expect(htmlencode(metadata(csdc.info(docxml, nil)).to_s)
-      .gsub(", :", ",\n:")).to be_equivalent_to output
-  end
-
-  it "uses new logo for invalid data" do
-    csdc = IsoDoc::Ogc::HtmlConvert.new({})
-    input = <<~INPUT
-      <ogc-standard xmlns="https://standards.opengeospatial.org/document">
-      <bibdata type="standard">
-        <title language="en" format="text/plain">Main Title</title>
-        <docidentifier type="ogc-external">http://www.example2.com</docidentifier>
-        <docidentifier type="ogc-internal">1000</docidentifier>
-        <docnumber>1000</docnumber>
-        <date type="published">
-          <on>yyyy-mm-dd</on>
-        </date>
-      </bibdata>
-      </ogc-standard>
-    INPUT
-    output = <<~OUTPUT
-      {:accesseddate=>"XXX",
-      :adapteddate=>"XXX",
-      :announceddate=>"XXX",
-      :circulateddate=>"XXX",
-      :confirmeddate=>"XXX",
-      :copieddate=>"XXX",
-      :copyright_holder=>"Open Geospatial Consortium",
-      :correcteddate=>"XXX",
-      :createddate=>"XXX",
-      :doclanguage=>"English",
-      :docnumber=>"1000",
-      :docnumeric=>"1000",
-      :doctitle=>"Main Title",
-      :externalid=>"http://www.example2.com",
-      :implementeddate=>"XXX",
-      :issueddate=>"XXX",
-      :lang=>"en",
-      :logo_new=>"#{File.join(logoloc, 'logo.2021.svg')}",
-      :logo_old=>"#{File.join(logoloc, 'logo.png')}",
-      :logo_word=>"#{File.join(logoloc, 'logo.2021.svg')}",
-      :obsoleteddate=>"XXX",
-      :publisheddate=>"yyyy-mm-dd",
-      :receiveddate=>"XXX",
-      :script=>"Latn",
-      :stable_untildate=>"XXX",
-      :transmitteddate=>"XXX",
-      :unchangeddate=>"XXX",
-      :unpublished=>true,
-      :updateddate=>"XXX",
-      :vote_endeddate=>"XXX",
-      :vote_starteddate=>"XXX"}
-    OUTPUT
-    docxml, = csdc.convert_init(input, "test", true)
-    expect(htmlencode(metadata(csdc.info(docxml, nil)).to_s)
-      .gsub(", :", ",\n:")).to be_equivalent_to output
-  end
-
-  it "processes contributor" do
-    input = <<~"INPUT"
-      <ogc-standard xmlns="#{Metanorma::Ogc::DOCUMENT_NAMESPACE}">
-      <bibdata>
-      <status>
-      <stage>approved</stage>
-      </status>
-              <contributor>
-          <role type="editor"/>
-          <person>
-            <name>
-              <completename>Fred Flintstone</completename>
-            </name>
-          </person>
-        </contributor>
-        <contributor>
-          <role type="author"/>
-          <person>
-            <name>
-              <forename>Barney</forename>
-              <surname>Rubble</surname>
-            </name>
-          </person>
-        </contributor>
-        <contributor>
-          <role type="contributor"/>
-          <person>
-            <name>
-              <forename>Pearl</forename>
-              <surname>Slaghoople</surname>
-            </name>
-          </person>
-        </contributor>
-      <ext>
-      <doctype>community-practice</doctype>
-      </ext>
-      </bibdata>
-      <sections/>
-      </ogc-standard>
-    INPUT
-
-    presxml = <<~OUTPUT
-      <ogc-standard xmlns='https://standards.opengeospatial.org/document' type='presentation'>
-           <bibdata>
-             <status>
-               <stage language=''>published</stage>
-               <stage language='en'>Published</stage>
-             </status>
-      <contributor>
-        <role type='editor'/>
-        <person>
-          <name>
-            <completename>Fred Flintstone</completename>
-          </name>
-        </person>
-      </contributor>
-      <contributor>
-        <role type='author'/>
-        <person>
-          <name>
-            <forename>Barney</forename>
-            <surname>Rubble</surname>
-          </name>
-        </person>
-      </contributor>
-      <contributor>
-        <role type='contributor'/>
-        <person>
-          <name>
-            <forename>Pearl</forename>
-            <surname>Slaghoople</surname>
-          </name>
-        </person>
-      </contributor>
-             <ext>
-               <doctype>community-practice</doctype>
-             </ext>
-           </bibdata>
-           <preface>
-               <clause type="toc" id="_" displayorder="1">
-            <title depth="1">Contents</title>
-          </clause>
-           </preface>
-          <sections/>
-      </ogc-standard>
-    OUTPUT
-    xml = Nokogiri::XML(IsoDoc::Ogc::PresentationXMLConvert.new(presxml_options)
-          .convert("test", input, true))
-    xml.at("//xmlns:localized-strings").remove
-    xml.at("//xmlns:metanorma-extension").remove
-    expect(xmlpp(strip_guid(xml.to_xml)))
-      .to be_equivalent_to xmlpp(presxml)
-  end
-
   it "orders terms in normal documents" do
     input = <<~"INPUT"
       <ogc-standard xmlns="#{Metanorma::Ogc::DOCUMENT_NAMESPACE}">
@@ -491,124 +109,6 @@ RSpec.describe IsoDoc::Ogc do
       .to be_equivalent_to xmlpp(presxml)
   end
 
-  it "changes approved to published" do
-    input = <<~"INPUT"
-      <ogc-standard xmlns="#{Metanorma::Ogc::DOCUMENT_NAMESPACE}">
-      <bibdata>
-      <status>
-      <stage>approved</stage>
-      </status>
-      <ext>
-      <doctype>community-practice</doctype>
-      </ext>
-      </bibdata>
-      <sections/>
-      </ogc-standard>
-    INPUT
-
-    presxml = <<~OUTPUT
-      <ogc-standard xmlns='https://standards.opengeospatial.org/document' type='presentation'>
-         <bibdata>
-           <status>
-             <stage language=''>published</stage>
-             <stage language='en'>Published</stage>
-           </status>
-           <ext>
-             <doctype>community-practice</doctype>
-           </ext>
-         </bibdata>
-           <preface> <clause type="toc" id="_" displayorder="1"> <title depth="1">Contents</title> </clause> </preface>
-         <sections/>
-       </ogc-standard>
-    OUTPUT
-    xml = Nokogiri::XML(IsoDoc::Ogc::PresentationXMLConvert.new(presxml_options)
-          .convert("test", input, true))
-    xml.at("//xmlns:localized-strings").remove
-    xml.at("//xmlns:metanorma-extension").remove
-    expect(xmlpp(strip_guid(xml.to_xml)))
-      .to be_equivalent_to xmlpp(presxml)
-  end
-
-  it "renders white-paper as technical-paper after cutoff date" do
-    input = <<~"INPUT"
-      <ogc-standard xmlns="#{Metanorma::Ogc::DOCUMENT_NAMESPACE}">
-      <bibdata>
-      <date type="published">2021-12-16</date>
-      <status>
-      <stage>approved</stage>
-      </status>
-      <ext>
-      <doctype>white-paper</doctype>
-      </ext>
-      </bibdata>
-      <sections/>
-      </ogc-standard>
-    INPUT
-
-    presxml = <<~OUTPUT
-      <ogc-standard xmlns='https://standards.opengeospatial.org/document' type='presentation'>
-         <bibdata>
-           <date type="published">2021-12-16</date>
-           <status>
-             <stage language=''>published</stage>
-             <stage language='en'>Published</stage>
-           </status>
-           <ext>
-             <doctype>technical-paper</doctype>
-           </ext>
-         </bibdata>
-           <preface> <clause type="toc" id="_" displayorder="1"> <title depth="1">Contents</title> </clause> </preface>
-         <sections/>
-       </ogc-standard>
-    OUTPUT
-    xml = Nokogiri::XML(IsoDoc::Ogc::PresentationXMLConvert.new(presxml_options)
-          .convert("test", input, true))
-    xml.at("//xmlns:localized-strings").remove
-    xml.at("//xmlns:metanorma-extension").remove
-    expect(xmlpp(strip_guid(xml.to_xml)))
-      .to be_equivalent_to xmlpp(presxml)
-  end
-
-  it "renders white-paper as white-paper before cutoff date" do
-    input = <<~"INPUT"
-      <ogc-standard xmlns="#{Metanorma::Ogc::DOCUMENT_NAMESPACE}">
-      <bibdata>
-      <date type="issued">2021-12-15</date>
-      <status>
-      <stage>approved</stage>
-      </status>
-      <ext>
-      <doctype>white-paper</doctype>
-      </ext>
-      </bibdata>
-      <sections/>
-      </ogc-standard>
-    INPUT
-
-    presxml = <<~OUTPUT
-      <ogc-standard xmlns='https://standards.opengeospatial.org/document' type='presentation'>
-         <bibdata>
-           <date type="issued">2021-12-15</date>
-           <status>
-             <stage language=''>published</stage>
-             <stage language='en'>Published</stage>
-           </status>
-           <ext>
-             <doctype>white-paper</doctype>
-           </ext>
-         </bibdata>
-           <preface> <clause type="toc" id="_" displayorder="1"> <title depth="1">Contents</title> </clause> </preface>
-         <sections/>
-       </ogc-standard>
-    OUTPUT
-    xml = Nokogiri::XML(IsoDoc::Ogc::PresentationXMLConvert.new(presxml_options)
-          .convert("test", input, true))
-    xml.at("//xmlns:localized-strings").remove
-    xml.at("//xmlns:metanorma-extension").remove
-    expect(xmlpp(strip_guid(xml.to_xml)))
-      .to be_equivalent_to xmlpp(presxml)
-  end
-
   it "processes pre" do
     input = <<~"INPUT"
       <ogc-standard xmlns="#{Metanorma::Ogc::DOCUMENT_NAMESPACE}">
@@ -637,362 +137,6 @@ RSpec.describe IsoDoc::Ogc do
              .gsub(%r{^.*<body}m, "<body")
              .gsub(%r{</body>.*}m, "</body>"),
            )).to be_equivalent_to output
-  end
-
-  it "processes keyword with no preface" do
-    input = <<~"INPUT"
-      <ogc-standard xmlns="#{Metanorma::Ogc::DOCUMENT_NAMESPACE}">
-      <bibdata>
-      <keyword>ABC</keyword>
-      <keyword>DEF</keyword>
-      </bibdata>
-      <sections/>
-      </ogc-standard>
-    INPUT
-
-    presxml = <<~OUTPUT
-      <ogc-standard xmlns='https://standards.opengeospatial.org/document' type="presentation">
-           <bibdata>
-             <keyword>ABC</keyword>
-             <keyword>DEF</keyword>
-           </bibdata>
-           <preface>
-               <clause type="toc" id="_" displayorder="1"> <title depth="1">Contents</title> </clause>
-             <clause id="_" type='keywords' displayorder="2">
-               <title depth='1'>I.<tab/>Keywords</title>
-               <p>The following are keywords to be used by search engines and document catalogues.</p>
-               <p>ABC, DEF</p>
-             </clause>
-           </preface>
-           <sections/>
-         </ogc-standard>
-    OUTPUT
-
-    output = <<~"OUTPUT"
-            #{HTML_HDR}
-            <div class="Section3" id="_">
-            <h1 class="IntroTitle">I.&#160; Keywords</h1>
-            <p>The following are keywords to be used by search engines and document catalogues.</p>
-            <p>ABC, DEF</p>
-          </div>
-        </div>
-      </body>
-    OUTPUT
-    xml = Nokogiri::XML(IsoDoc::Ogc::PresentationXMLConvert.new(presxml_options)
-          .convert("test", input, true))
-    xml.at("//xmlns:localized-strings").remove
-    xml.at("//xmlns:metanorma-extension").remove
-    expect(xmlpp(strip_guid(xml.to_xml)))
-      .to be_equivalent_to xmlpp(presxml)
-    expect(xmlpp(
-             IsoDoc::Ogc::HtmlConvert.new({})
-             .convert("test", presxml, true)
-             .gsub(%r{^.*<body}m, "<body")
-             .gsub(%r{</body>.*}m, "</body>"),
-           )).to be_equivalent_to xmlpp(output)
-  end
-
-  it "processes keyword with preface" do
-    input = <<~"INPUT"
-      <ogc-standard xmlns="#{Metanorma::Ogc::DOCUMENT_NAMESPACE}">
-      <bibdata>
-      <keyword>ABC</keyword>
-      <keyword>DEF</keyword>
-      </bibdata>
-      <preface>
-      <abstract id="A"/>
-      </preface>
-      <sections/>
-      </ogc-standard>
-    INPUT
-
-    presxml = <<~OUTPUT
-          <ogc-standard xmlns='https://standards.opengeospatial.org/document' type="presentation">
-               <bibdata>
-                 <keyword>ABC</keyword>
-                 <keyword>DEF</keyword>
-               </bibdata>
-               <preface>
-                   <clause type="toc" id="_" displayorder="1"> <title depth="1">Contents</title> </clause>
-               <abstract id='A' displayorder="2">
-        <title>I.</title>
-      </abstract>
-                 <clause id="_" type='keywords' displayorder="3">
-                   <title depth='1'>II.<tab/>Keywords</title>
-                   <p>The following are keywords to be used by search engines and document catalogues.</p>
-                   <p>ABC, DEF</p>
-                 </clause>
-               </preface>
-               <sections/>
-             </ogc-standard>
-    OUTPUT
-
-    output = <<~"OUTPUT"
-       #{HTML_HDR}
-          <br/>
-          <div id='A'>
-            <h1 class='AbstractTitle'>I.</h1>
-          </div>
-          <div class='Section3' id='_'>
-            <h1 class='IntroTitle'>II.&#160; Keywords</h1>
-            <p>
-              The following are keywords to be used by search engines and document
-              catalogues.
-            </p>
-            <p>ABC, DEF</p>
-          </div>
-        </div>
-      </body>
-    OUTPUT
-    xml = Nokogiri::XML(IsoDoc::Ogc::PresentationXMLConvert.new(presxml_options)
-          .convert("test", input, true))
-    xml.at("//xmlns:localized-strings").remove
-    xml.at("//xmlns:metanorma-extension").remove
-    expect(xmlpp(strip_guid(xml.to_xml)))
-      .to be_equivalent_to xmlpp(presxml)
-    expect(xmlpp(
-             IsoDoc::Ogc::HtmlConvert.new({})
-             .convert("test", presxml, true)
-             .gsub(%r{^.*<body}m, "<body")
-             .gsub(%r{</body>.*}m, "</body>"),
-           )).to be_equivalent_to xmlpp(output)
-  end
-
-  it "processes submitting organisations with no preface" do
-    input = <<~"INPUT"
-      <ogc-standard xmlns="#{Metanorma::Ogc::DOCUMENT_NAMESPACE}">
-       <bibdata>
-               <contributor>
-                 <role type="author"/>
-                 <organization>
-                   <name>OGC</name>
-                 </organization>
-               </contributor>
-               <contributor>
-                 <role type="author"/>
-                 <organization>
-                   <name>DEF</name>
-                 </organization>
-               </contributor>
-            </bibdata>
-      <sections/>
-      </ogc-standard>
-    INPUT
-
-    presxml = <<~OUTPUT
-          <ogc-standard xmlns='https://standards.opengeospatial.org/document' type="presentation">
-          <bibdata>
-               <contributor>
-                 <role type="author"/>
-                 <organization>
-                   <name>OGC</name>
-                 </organization>
-               </contributor>
-               <contributor>
-                 <role type="author"/>
-                 <organization>
-                   <name>DEF</name>
-                 </organization>
-               </contributor>
-            </bibdata>
-            <preface>
-                <clause type="toc" id="_" displayorder="1"> <title depth="1">Contents</title> </clause>
-        <clause id='_' type='submitting_orgs' displayorder="2">
-          <title depth='1'>I.<tab/>Submitting Organizations</title>
-          <p>The following organizations submitted this Document to the Open
-            Geospatial Consortium (OGC):</p>
-          <ul>
-            <li>OGC</li>
-            <li>DEF</li>
-          </ul>
-        </clause>
-            </preface>
-      <sections/>
-      </ogc-standard>
-    OUTPUT
-
-    output = <<~"OUTPUT"
-              #{HTML_HDR}
-              <div class="Section3" id="_">
-              <h1 class="IntroTitle">I.&#160; Submitting Organizations</h1>
-              <p>The following organizations submitted this Document to the Open
-      Geospatial Consortium (OGC):</p>
-      <ul>
-        <li>OGC</li>
-        <li>DEF</li>
-      </ul>
-            </div>
-          </div>
-        </body>
-    OUTPUT
-    xml = Nokogiri::XML(IsoDoc::Ogc::PresentationXMLConvert.new(presxml_options)
-          .convert("test", input, true))
-    xml.at("//xmlns:localized-strings").remove
-    xml.at("//xmlns:metanorma-extension").remove
-    expect(xmlpp(strip_guid(xml.to_xml)))
-      .to be_equivalent_to xmlpp(presxml)
-    expect(xmlpp(IsoDoc::Ogc::HtmlConvert.new({})
-      .convert("test", presxml, true)
-      .gsub(%r{^.*<body}m, "<body").gsub(%r{</body>.*}m, "</body>")))
-      .to be_equivalent_to xmlpp(output)
-  end
-
-  it "processes submitting organisations with preface" do
-    input = <<~"INPUT"
-      <ogc-standard xmlns="#{Metanorma::Ogc::DOCUMENT_NAMESPACE}">
-       <bibdata>
-               <contributor>
-                 <role type="author"/>
-                 <organization>
-                   <name>OGC</name>
-                 </organization>
-               </contributor>
-               <contributor>
-                 <role type="author"/>
-                 <organization>
-                   <name>DEF</name>
-                 </organization>
-               </contributor>
-            </bibdata>
-      <preface>
-      <abstract id="A"/>
-      <clause type="security" id="B"/>
-      </preface>
-      <sections/>
-      </ogc-standard>
-    INPUT
-
-    presxml = <<~OUTPUT
-          <ogc-standard xmlns='https://standards.opengeospatial.org/document' type="presentation">
-          <bibdata>
-               <contributor>
-                 <role type="author"/>
-                 <organization>
-                   <name>OGC</name>
-                 </organization>
-               </contributor>
-               <contributor>
-                 <role type="author"/>
-                 <organization>
-                   <name>DEF</name>
-                 </organization>
-               </contributor>
-            </bibdata>
-            <preface>
-                <clause type="toc" id="_" displayorder="1"> <title depth="1">Contents</title> </clause>
-                 <abstract id='A' displayorder="2">
-                   <title>I.</title>
-                 </abstract>
-                 <clause type='security' id='B' displayorder="3">
-        <title>II.</title>
-      </clause>
-                 <clause id='_' type='submitting_orgs' displayorder="4">
-                   <title depth='1'>
-                     III.
-                     <tab/>
-                     Submitting Organizations
-                   </title>
-                   <p>The following organizations submitted this Document to the Open
-      Geospatial Consortium (OGC):</p>
-                   <ul>
-                     <li>OGC</li>
-                     <li>DEF</li>
-                   </ul>
-                 </clause>
-               </preface>
-      <sections/>
-      </ogc-standard>
-    OUTPUT
-
-    output = <<~OUTPUT
-          <body lang='EN-US' xml:lang='EN-US' link='blue' vlink='#954F72' class='container'>
-        <div class='title-section'>
-          <p>&#160;</p>
-        </div>
-        <br/>
-        <div class='prefatory-section'>
-          <p>&#160;</p>
-        </div>
-        <br/>
-        <div class='main-section'>
-          <br/>
-          <div id="_" class="TOC">
-          <h1 class="IntroTitle">Contents</h1>
-        </div>
-        <br/>
-          <div id='A'>
-            <h1 class='AbstractTitle'>I.</h1>
-          </div>
-          <div class='Section3' id='B'>
-        <h1 class='IntroTitle'>II.</h1>
-      </div>
-          <div class='Section3' id='_'>
-            <h1 class='IntroTitle'> III. &#160; Submitting Organizations </h1>
-            <p>
-              The following organizations submitted this Document to the Open
-              Geospatial Consortium (OGC):
-            </p>
-            <ul>
-              <li>OGC</li>
-              <li>DEF</li>
-            </ul>
-          </div>
-        </div>
-      </body>
-    OUTPUT
-    xml = Nokogiri::XML(IsoDoc::Ogc::PresentationXMLConvert.new(presxml_options)
-          .convert("test", input, true))
-    xml.at("//xmlns:localized-strings").remove
-    xml.at("//xmlns:metanorma-extension").remove
-    expect(xmlpp(strip_guid(xml.to_xml)))
-      .to be_equivalent_to xmlpp(presxml)
-    expect(xmlpp(IsoDoc::Ogc::HtmlConvert.new({})
-      .convert("test", presxml, true)
-      .gsub(%r{^.*<body}m, "<body").gsub(%r{</body>.*}m, "</body>")))
-      .to be_equivalent_to xmlpp(output)
-  end
-
-  it "processes keyword and abstract in HTML head" do
-    presxml = <<~OUTPUT
-          <ogc-standard xmlns='https://standards.opengeospatial.org/document' type="presentation">
-               <bibdata>
-                 <keyword>ABC</keyword>
-                 <keyword>DEF</keyword>
-                 <keyword>"Double &#x3c;quote&#x3e;"</keyword>
-                 <keyword>'Single quote'</keyword>
-                 <abstract><p>This is a &#x3c;description&#x3e;.</p>
-              <quote>This is a <em>blockquote</em> within a description. "Double quote" and 'Single quote'.</quote>
-                 </abstract>
-               </bibdata>
-               <preface>
-                <clause type="toc" id="_" displayorder="1"> <title depth="1">Contents</title> </clause>
-               <abstract id='A' displayorder="2">
-        <title>I.</title>
-      </abstract>
-                 <clause id="_" type='keywords' displayorder="3">
-                   <title depth='1'>II.<tab/>Keywords</title>
-                   <p>The following are keywords to be used by search engines and document catalogues.</p>
-                   <p>ABC, DEF</p>
-                 </clause>
-               </preface>
-               <sections/>
-             </ogc-standard>
-    OUTPUT
-
-    output = <<~OUTPUT
-      <html>
-         <meta name="keywords" content="ABC, DEF, &quot;Double &lt;quote&gt;&quot;, 'Single quote'"/>
-         <meta name="description" content="This is a &lt;description&gt;. This is a blockquote within a description. &quot;Double quote&quot; and 'Single quote'."/>
-      </html>
-    OUTPUT
-
-    FileUtils.rm_f("test.html")
-    IsoDoc::Ogc::HtmlConvert.new({ filename: "test" })
-      .convert("test", presxml, false)
-    doc = Nokogiri::XML(File.read("test.html"))
-    out = doc.xpath("//head/meta[@name = 'keywords' or @name = 'description']")
-    expect(xmlpp("<html>#{out.to_xml}</html>"))
-      .to be_equivalent_to xmlpp(output)
   end
 
   it "processes simple terms & definitions" do
